+++ rqt_controller_manager/CHANGELOG.rst
@@ -2,6 +2,11 @@
 Changelog for package rqt_controller_manager
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+4.3.0 (2024-01-20)
+------------------
+* Fix rqt controller manager crash on ros2_control restart (`#1273 <https://github.com/ros-controls/ros2_control/issues/1273>`_)
+* Contributors: Sai Kishor Kothakota
+
 4.2.0 (2023-12-12)
 ------------------